+++ ComposeSimpleApp/app/src/main/java/com/example/myapplication/model/room/PostRepository.kt
@@ -4,8 +4,6 @@ import com.example.myapplication.model.PostEntity
 
 class PostRepository(private val dao: PostDao) {
 
-    val getAllPosts = dao.getAllPosts()
-
     suspend fun addPost(postEntity: PostEntity) {
         dao.insert(postEntity)
     }
