+++ ComposeSimpleApp/app/src/main/java/com/example/myapplication/ui/BaseComposableFragment.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example.myapplication.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
