+++ ComposeSimpleApp/app/src/main/java/com/example/myapplication/ui/features/home/details/ItemDetailFragment.kt
@@ -2,21 +2,15 @@ package com.example.myapplication.ui.features.home.details
 
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.Button
-import androidx.compose.material.Card
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.myapplication.BaseComposableFragment
+import com.example.myapplication.ui.BaseComposableFragment
 import com.example.myapplication.R
 import com.example.myapplication.model.Post
 import com.example.myapplication.ui.InteractButton
