+++ ComposeSimpleApp/app/src/main/java/com/example/myapplication/ui/features/profile/ProfileFragment.kt
@@ -1,26 +1,15 @@
 package com.example.myapplication.ui.features.profile
 
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.Card
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.*
-import com.example.myapplication.BaseComposableFragment
-import com.example.myapplication.R
+import com.example.myapplication.ui.BaseComposableFragment
 import com.example.myapplication.model.User
 import com.example.myapplication.ui.ProfileImage
 import com.example.myapplication.ui.Statistic
